+++ primitives/Cargo.toml
@@ -16,18 +16,28 @@ exclude = ["tests", "contrib"]
 
 [features]
 default = ["std"]
-std = ["alloc", "internals/std"]
-alloc = ["internals/alloc"]
-serde = ["actual-serde", "internals/serde"]
+std = ["alloc", "internals/std", "io/std", "units/std"]
+alloc = ["internals/alloc", "io/alloc", "units/alloc"]
+serde = ["actual-serde", "internals/serde", "units/serde"]
 
 [dependencies]
 internals = { package = "bitcoin-internals", version = "0.3.0" }
+io = { package = "bitcoin-io", version = "0.1.1", default-features = false }
+units = { package = "bitcoin-units", version = "0.1.0", default-features = false }
+
+ordered = { version = "0.2.0", optional = true }
 
 # Do NOT use this as a feature! Use the `serde` feature instead.
-actual-serde = { package = "serde", version = "1.0.103", default-features = false, optional = true }
+actual-serde = { package = "serde", version = "1.0.103", default-features = false, features = ["derive"], optional = true }
 
 [dev-dependencies]
+
+[target.'cfg(mutate)'.dev-dependencies]
+mutagen = { git = "https://github.com/llogiq/mutagen" }
 
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
+
+[lints.rust]
+unexpected_cfgs = { level = "deny", check-cfg = ['cfg(mutate)'] }
